planner: move PhysicalIndexScan related logic into physical op dir
