+++ pkg/planner/core/explain.go
@@ -20,120 +20,10 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/pingcap/tidb/pkg/expression"
 	"github.com/pingcap/tidb/pkg/planner/property"
 	"github.com/pingcap/tidb/pkg/sessionctx/vardef"
-	"github.com/pingcap/tidb/pkg/statistics"
-	"github.com/pingcap/tidb/pkg/util/plancodec"
-	"github.com/pingcap/tidb/pkg/util/stringutil"
 	"github.com/pingcap/tipb/go-tipb"
 )
-
-// ExplainID overrides the ExplainID in order to match different range.
-func (p *PhysicalIndexScan) ExplainID(_ ...bool) fmt.Stringer {
-	return stringutil.MemoizeStr(func() string {
-		if p.SCtx() != nil && p.SCtx().GetSessionVars().StmtCtx.IgnoreExplainIDSuffix {
-			return p.TP()
-		}
-		return p.TP() + "_" + strconv.Itoa(p.ID())
-	})
-}
-
-// TP overrides the TP in order to match different range.
-func (p *PhysicalIndexScan) TP(_ ...bool) string {
-	if p.isFullScan() {
-		return plancodec.TypeIndexFullScan
-	}
-	return plancodec.TypeIndexRangeScan
-}
-
-// ExplainInfo implements Plan interface.
-func (p *PhysicalIndexScan) ExplainInfo() string {
-	return p.AccessObject().String() + ", " + p.OperatorInfo(false)
-}
-
-// ExplainNormalizedInfo implements Plan interface.
-func (p *PhysicalIndexScan) ExplainNormalizedInfo() string {
-	return p.AccessObject().NormalizedString() + ", " + p.OperatorInfo(true)
-}
-
-// OperatorInfo implements DataAccesser interface.
-func (p *PhysicalIndexScan) OperatorInfo(normalized bool) string {
-	ectx := p.SCtx().GetExprCtx().GetEvalCtx()
-	redact := p.SCtx().GetSessionVars().EnableRedactLog
-	var buffer strings.Builder
-	if len(p.rangeInfo) > 0 {
-		if !normalized {
-			buffer.WriteString("range: decided by ")
-			buffer.WriteString(p.rangeInfo)
-			buffer.WriteString(", ")
-		}
-	} else if p.haveCorCol() {
-		if normalized {
-			buffer.WriteString("range: decided by ")
-			buffer.Write(expression.SortedExplainNormalizedExpressionList(p.AccessCondition))
-			buffer.WriteString(", ")
-		} else {
-			buffer.WriteString("range: decided by [")
-			for i, expr := range p.AccessCondition {
-				if i != 0 {
-					buffer.WriteString(" ")
-				}
-				buffer.WriteString(expr.StringWithCtx(ectx, redact))
-			}
-			buffer.WriteString("], ")
-		}
-	} else if len(p.Ranges) > 0 {
-		if normalized {
-			buffer.WriteString("range:[?,?], ")
-		} else if !p.isFullScan() {
-			buffer.WriteString("range:")
-			for _, idxRange := range p.Ranges {
-				buffer.WriteString(idxRange.Redact(redact))
-				buffer.WriteString(", ")
-			}
-		}
-	}
-	buffer.WriteString("keep order:")
-	buffer.WriteString(strconv.FormatBool(p.KeepOrder))
-	if p.Desc {
-		buffer.WriteString(", desc")
-	}
-	if !normalized {
-		if p.usedStatsInfo != nil {
-			str := p.usedStatsInfo.FormatForExplain()
-			if len(str) > 0 {
-				buffer.WriteString(", ")
-				buffer.WriteString(str)
-			}
-		} else if p.StatsInfo().StatsVersion == statistics.PseudoVersion {
-			// This branch is not needed in fact, we add this to prevent test result changes under planner/cascades/
-			buffer.WriteString(", stats:pseudo")
-		}
-	}
-	return buffer.String()
-}
-
-func (p *PhysicalIndexScan) haveCorCol() bool {
-	for _, cond := range p.AccessCondition {
-		if len(expression.ExtractCorColumns(cond)) > 0 {
-			return true
-		}
-	}
-	return false
-}
-
-func (p *PhysicalIndexScan) isFullScan() bool {
-	if len(p.rangeInfo) > 0 || p.haveCorCol() {
-		return false
-	}
-	for _, ran := range p.Ranges {
-		if !ran.IsFullRange(false) {
-			return false
-		}
-	}
-	return true
-}
 
 // ExplainInfo implements Plan interface.
 func (p *PhysicalTableReader) ExplainInfo() string {
