+++ pkg/planner/core/operator/physicalop/BUILD.bazel
@@ -13,6 +13,7 @@ go_library(
         "physical_hash_join.go",
         "physical_index_hash_join.go",
         "physical_index_join.go",
+        "physical_index_scan.go",
         "physical_limit.go",
         "physical_lock.go",
         "physical_max_one_row.go",
