+++ pkg/planner/core/plan_to_pb.go
@@ -21,38 +21,8 @@ import (
 	"github.com/pingcap/tidb/pkg/meta/model"
 	"github.com/pingcap/tidb/pkg/planner/core/base"
 	"github.com/pingcap/tidb/pkg/planner/core/operator/physicalop"
-	"github.com/pingcap/tidb/pkg/table/tables"
-	"github.com/pingcap/tidb/pkg/util"
-	"github.com/pingcap/tidb/pkg/util/ranger"
 	"github.com/pingcap/tipb/go-tipb"
 )
-
-// checkCoverIndex checks whether we can pass unique info to TiKV. We should push it if and only if the length of
-// range and index are equal.
-func checkCoverIndex(idx *model.IndexInfo, ranges []*ranger.Range) bool {
-	// If the index is (c1, c2) but the query range only contains c1, it is not a unique get.
-	if !idx.Unique {
-		return false
-	}
-	for _, rg := range ranges {
-		if len(rg.LowVal) != len(idx.Columns) {
-			return false
-		}
-		for _, v := range rg.LowVal {
-			if v.IsNull() {
-				// a unique index may have duplicated rows with NULLs, so we cannot set the unique attribute to true when the range has NULL
-				// please see https://github.com/pingcap/tidb/issues/29650 for more details
-				return false
-			}
-		}
-		for _, v := range rg.HighVal {
-			if v.IsNull() {
-				return false
-			}
-		}
-	}
-	return true
-}
 
 // FindColumnInfoByID finds ColumnInfo in cols by ID.
 func FindColumnInfoByID(colInfos []*model.ColumnInfo, id int64) *model.ColumnInfo {
@@ -176,38 +146,6 @@ func (e *PhysicalExchangeReceiver) ToPB(ctx *base.BuildPBContext, _ kv.StoreType
 		FineGrainedShuffleStreamCount: e.TiFlashFineGrainedShuffleStreamCount,
 		FineGrainedShuffleBatchSize:   ctx.TiFlashFineGrainedShuffleBatchSize,
 	}, nil
-}
-
-// ToPB implements PhysicalPlan ToPB interface.
-func (p *PhysicalIndexScan) ToPB(_ *base.BuildPBContext, _ kv.StoreType) (*tipb.Executor, error) {
-	columns := make([]*model.ColumnInfo, 0, p.Schema().Len())
-	tableColumns := p.Table.Cols()
-	for _, col := range p.Schema().Columns {
-		if col.ID == model.ExtraHandleID {
-			columns = append(columns, model.NewExtraHandleColInfo())
-		} else if col.ID == model.ExtraPhysTblID {
-			columns = append(columns, model.NewExtraPhysTblIDColInfo())
-		} else {
-			columns = append(columns, FindColumnInfoByID(tableColumns, col.ID))
-		}
-	}
-	var pkColIDs []int64
-	if p.NeedCommonHandle {
-		pkColIDs = tables.TryGetCommonPkColumnIds(p.Table)
-	}
-	idxExec := &tipb.IndexScan{
-		TableId:          p.Table.ID,
-		IndexId:          p.Index.ID,
-		Columns:          util.ColumnsToProto(columns, p.Table.PKIsHandle, true, false),
-		Desc:             p.Desc,
-		PrimaryColumnIds: pkColIDs,
-	}
-	if p.isPartition {
-		idxExec.TableId = p.physicalTableID
-	}
-	unique := checkCoverIndex(p.Index, p.Ranges)
-	idxExec.Unique = &unique
-	return &tipb.Executor{Tp: tipb.ExecType_TypeIndexScan, IdxScan: idxExec}, nil
 }
 
 // toPB4PhysicalSort implements PhysicalPlan ToPB interface.
