+++ pkg/planner/util/utilfuncp/BUILD.bazel
@@ -8,10 +8,14 @@ go_library(
     deps = [
         "//pkg/expression",
         "//pkg/kv",
+        "//pkg/meta/model",
+        "//pkg/parser/ast",
         "//pkg/planner/core/base",
         "//pkg/planner/property",
         "//pkg/planner/util/costusage",
         "//pkg/planner/util/optimizetrace",
+        "//pkg/table",
+        "//pkg/types",
         "//pkg/util/execdetails",
         "@com_github_pingcap_tipb//go-tipb",
     ],
